+++ src/lib/gpioEx/mcp23008/Mcp23008.h
@@ -22,7 +22,7 @@ class GpioMcp23008 : public Gpio {
   private:
     bool found = false;
 
-    int mode[8] = { INPUT, INPUT, INPUT, INPUT, INPUT, INPUT, INPUT, INPUT };
+    int8_t mode[8] = { INPUT, INPUT, INPUT, INPUT, INPUT, INPUT, INPUT, INPUT };
     bool state[8] = { false, false, false, false, false, false, false, false };
 };
