+++ src/lib/gpioEx/ssr74HC595/Ssr74HC595.h
@@ -24,7 +24,7 @@ class GpioSsr74HC595 : public Gpio {
     uint32_t register_value = 0;
     bool found = false;
 
-    int mode[32];
+    int8_t mode[32];
     bool state[32];
 };
